Backend/requirements.txt
firebase==4.0.1
firebase-admin==6.2.0
Flask==3.0.0
ultralytics==8.0.203
timm==0.9.8
selenium==4.14.0
requests==2.29.0
Pyrebase4==4.7.1
Pillow==10.1.0
pandas
opencv-python
numpy
PyJWT
python-dotenv

